+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.1",
     "bootstrap": "^4.6.0",
+    "chart.js": "^3.3.2",
     "cors": "^2.8.5",
     "d3": "^6.7.0",
     "enzyme": "^3.11.0",
@@ -16,6 +17,7 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.0",
+    "react-chartjs-2": "^3.0.3",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
