+++ src/styles/index.css
@@ -164,7 +164,7 @@ nav div {
 
 .bar-chart {
   display: flex;
-  border: 1px solid black;
+  /* border: 1px solid black; */
   padding: 10px;
   margin-bottom: 15px;
 }
